pyproject.toml
[project]
name = "aint"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13,<3.14"
dependencies = [
    "adaptix>=3.0.0b11",
    "click>=8.3.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.0",
    "openai>=2.14.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
aint = "aint.cli_entrypoint:main"

[build-system]
requires = ["setuptools >= 80.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["aint"]
